+++ src/host/coreaudio/macos/property_listener.rs
@@ -11,7 +11,7 @@ use crate::BuildStreamError;
 
 /// A double-indirection to be able to pass a closure (a fat pointer)
 /// via a single c_void.
-struct PropertyListenerCallbackWrapper(Box<dyn FnMut()>);
+struct PropertyListenerCallbackWrapper(Box<dyn FnMut() + Send>);
 
 /// Maintain an audio object property listener.
 /// The listener will be removed when this type is dropped.
@@ -24,7 +24,7 @@ pub struct AudioObjectPropertyListener {
 
 impl AudioObjectPropertyListener {
     /// Attach the provided callback as a audio object property listener.
-    pub fn new<F: FnMut() + 'static>(
+    pub fn new<F: FnMut() + Send + 'static>(
         audio_object_id: AudioObjectID,
         property_address: AudioObjectPropertyAddress,
         callback: F,
@@ -49,6 +49,7 @@ impl AudioObjectPropertyListener {
     /// Explicitly remove the property listener.
     /// Use this method if you need to explicitly handle failure to remove
     /// the property listener.
+    #[allow(dead_code)]
     pub fn remove(mut self) -> Result<(), BuildStreamError> {
         self.remove_inner()
     }